Makes more sense for this method to be in ChangeCalculator, not in MoneyUtil
