+++ src/main/java/tdd/vendingMachine/money/change/ChangeCalculator.java
@@ -3,22 +3,50 @@ package tdd.vendingMachine.money.change;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
+import lombok.Data;
 import org.joda.money.Money;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import tdd.vendingMachine.money.coin.entity.Coin;
+import tdd.vendingMachine.money.coin.factory.CoinFactory;
 
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
+import java.util.*;
 import java.util.stream.Collectors;
 
 public class ChangeCalculator {
 
-	public static Map<Coin, Integer> calculate(Map<Coin, Integer> availableCoins, Money requiredMoney) {
+	@Data
+	private static class Subset {
+		List<Integer> amounts;
+
+		Integer fromIndex;
+
+		Integer target;
+
+		Integer sumInStack;
+
+		Set<List<Integer>> solutions;
+
+		Stack<Integer> stack;
+
+		Subset(List<Integer> amounts, Integer target) {
+			this.amounts = amounts;
+			this.target = target;
+			solutions = Sets.newLinkedHashSet();
+			stack = new Stack<>();
+			fromIndex = sumInStack = 0;
+		}
+	}
+
+	public static Map<Coin, Integer> calculateChange(Map<Coin, Integer> availableCoins, Money requiredMoney) {
 		int requiredAmount = requiredMoney.getAmountMinorInt();
 		return chooseMostDispersedCoinSet(getAllSolutions(availableCoins, Lists.newArrayList(), requiredAmount));
+	}
+
+	public static Map<Coin, Integer> calculateChangeDifference(Map<Coin, Integer> storage, Money target) {
+		List<Coin> coins = toCoins(storage);
+		List<Integer> amounts = toAmounts(coins);
+		return getBestSubset(amounts, target);
 	}
 
 	private static Set<Map<Coin, Integer>> getAllSolutions(Map<Coin, Integer> availableCoins,
@@ -95,6 +123,89 @@ public class ChangeCalculator {
 
 	private static int getCoinValue(Coin coin) {
 		return coin.getNominal().getAmountMinorInt();
+	}
+
+	private static List<Coin> toCoins(Map<Coin, Integer> storage) {
+		List<Coin> coins = Lists.newArrayList();
+		storage.entrySet().forEach(consumer -> {
+			int i = 0;
+			while(i < consumer.getValue()) {
+				coins.add(consumer.getKey());
+				i++;
+			}
+		});
+		return coins;
+	}
+
+	private static List<Integer> toAmounts(List<Coin> coins) {
+		return coins
+			.stream()
+			.map(coin -> coin.getNominal().getAmountMinorInt())
+			.collect(Collectors.toList());
+	}
+
+	private static Map<Coin, Integer> getBestSubset(List<Integer> amounts, Money target) {
+		Map<Coin, Integer> subset = CoinFactory.emptyCoinStorage();
+
+		List<Integer> nominals = getSubset(amounts, target.getAmountMinorInt());
+
+		if (nominals == null) {
+			return null;
+		}
+
+		if (!nominals.isEmpty()) {
+			nominals.forEach(consumer -> {
+				Coin coin = CoinFactory.ofAmount(consumer);
+				subset.put(coin, subset.get(coin) + 1);
+			});
+		}
+
+		return subset;
+	}
+
+	// based on http://codereview.stackexchange.com/q/36214
+	private static List<Integer> getSubset(List<Integer> amounts, Integer target) {
+		Subset subset = new Subset(amounts, target);
+		getSubsets(subset);
+		return subset.getSolutions().stream()
+			.sorted((base, compare) -> Integer.compare(base.size(), compare.size()))
+			.findFirst().orElse(null);
+	}
+
+	private static void getSubsets(Subset subset) {
+		if (sumInStackEqualsTarget(subset)) {
+			addSolution(subset);
+		}
+
+		for (Integer currentIndex = subset.getFromIndex(); currentIndex < subset.getAmounts().size(); currentIndex++) {
+			if (sumInStackAndNextAmountIsNotGraterThanTarget(subset, currentIndex)) {
+				doIterateSubset(subset, currentIndex);
+			}
+		}
+	}
+
+	private static boolean sumInStackAndNextAmountIsNotGraterThanTarget(Subset subset, Integer currentIndex) {
+		return subset.getSumInStack() + subset.getAmounts().get(currentIndex) <= subset.getTarget();
+	}
+
+	private static boolean sumInStackEqualsTarget(Subset subset) {
+		return subset.getSumInStack().equals(subset.getTarget());
+	}
+
+	private static void addSolution(Subset subset) {
+		List<Integer> list = Lists.newArrayList();
+		subset.getStack().forEach(list::add);
+		subset.getSolutions().add(list);
+	}
+
+	private static void doIterateSubset(Subset subset, Integer currentIndex) {
+		subset.getStack().push(subset.getAmounts().get(currentIndex));
+		subset.setSumInStack(subset.getSumInStack() + subset.getAmounts().get(currentIndex));
+		subset.setFromIndex(currentIndex + 1);
+		if (subset.getSolutions().isEmpty()) {
+			getSubsets(subset);
+		}
+		subset.setSumInStack(subset.getSumInStack() - subset.getStack().pop());
 	}
 
 }
