+++ src/test/java/tdd/vendingMachine/money/util/MoneyUtilTest.java
@@ -5,7 +5,6 @@ import org.assertj.core.api.Assertions;
 import org.junit.Test;
 import tdd.vendingMachine.money.coin.entity.Coin;
 import tdd.vendingMachine.money.coin.factory.CoinFactory;
-import tdd.vendingMachine.money.factory.MoneyFactory;
 
 import java.util.Map;
 
@@ -67,49 +66,5 @@ public class MoneyUtilTest {
 		Assertions.assertThat(difference.get(CoinFactory.create020())).isEqualTo(4);
 		Assertions.assertThat(difference.get(CoinFactory.create050())).isEqualTo(1);
 	}
-
-	@Test
-	public void extracts_subset_of_money_for_money_valued_0_dot_40() {
-		Map<Coin, Integer> map1 = Maps.newLinkedHashMap();
-		map1.put(CoinFactory.create020(), 3);
-
-		Map<Coin, Integer> subset = MoneyUtil.subset(map1, MoneyFactory.of(.4));
-
-		Assertions.assertThat(subset).hasSize(6);
-		Assertions.assertThat(subset.get(CoinFactory.create020())).isEqualTo(2);
-	}
-
-	@Test
-	public void extracts_subset_of_money_for_money_valued_1_dot_70() {
-		Map<Coin, Integer> map1 = Maps.newLinkedHashMap();
-		map1.put(CoinFactory.create010(), 2);
-		map1.put(CoinFactory.create020(), 3);
-		map1.put(CoinFactory.create100(), 2);
-
-		Map<Coin, Integer> subset = MoneyUtil.subset(map1, MoneyFactory.of(1.7));
-
-		Assertions.assertThat(subset).hasSize(6);
-		Assertions.assertThat(subset.get(CoinFactory.create010())).isEqualTo(1);
-		Assertions.assertThat(subset.get(CoinFactory.create020())).isEqualTo(3);
-		Assertions.assertThat(subset.get(CoinFactory.create100())).isEqualTo(1);
-	}
-
-	@Test
-	public void extracts_subset_of_money_for_money_valued_3_dot_60() {
-		Map<Coin, Integer> map1 = Maps.newLinkedHashMap();
-		map1.put(CoinFactory.create010(), 3);
-		map1.put(CoinFactory.create020(), 3);
-		map1.put(CoinFactory.create050(), 3);
-		map1.put(CoinFactory.create100(), 3);
-
-		Map<Coin, Integer> subset = MoneyUtil.subset(map1, MoneyFactory.of(3.6));
-
-		Assertions.assertThat(subset).hasSize(6);
-		Assertions.assertThat(subset.get(CoinFactory.create010())).isEqualTo(2);
-		Assertions.assertThat(subset.get(CoinFactory.create020())).isEqualTo(2);
-		Assertions.assertThat(subset.get(CoinFactory.create050())).isEqualTo(2);
-		Assertions.assertThat(subset.get(CoinFactory.create100())).isEqualTo(2);
-	}
-
 
 }
